Port e2e tx_propagation tests
